myworld/app_prototipo/requirements.txt
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.0.0
Pillow==10.0.0
python-dateutil==2.8.2
openpyxl==3.1.2
reportlab==4.0.4
requests==2.31.0
django-environ==0.11.0
psycopg2-binary==2.9.7
celery==5.3.0
redis==4.6.0